renaming and rearrangement
